action-server: Move action server entry point to apps/
Change-type: patch
